Fix: navbar not showing on initial page load
