+++ src/routes/+layout.svelte
@@ -1,12 +1,8 @@
 <script>
 	import "../app.pcss";
 	import Navbar from "$lib/components/Navbar.svelte";
-	import { showNavbar } from "$lib/stores/navbarStore";
+	// import { showNavbar } from "$lib/stores/navbarStore";
 </script>
 
-{#await Promise.resolve() then _}
-	{#if showNavbar}
-		<Navbar />
-	{/if}
-{/await}
+<Navbar />
 <slot />
